Create file with environment variables and aliases
Create a file when the snap is installed and refreshed which exports
the JAVA_HOME environment variable and defines normal JDK aliases
for the 'openjdk.*' commands.

Add desktop interfaces to the 'java' command so that it can run Java
desktop applications.

Add the 'jarsigner' command.

Add a cleanup part called 'del' that deletes duplicate files already
in the 'core20' base snap.

Rewrite the README file to reflect the changes.