Add all YAML files to main branch for development
Merging changes from one branch into another often results in
conflicts and wipes out the correct branch-specific changes in the
'snapcraft.yaml' file.

Fix the Git workflow by putting all of the Snapcraft YAML files under
'snap/local' on the main branch and make 'snapcraft.yaml' a symbolic
link whose target depends on the branch. Develop and test the Snap
packages for all of the channels on the main branch by switching the
symbolic link. Merge changes only from the main branch into the other
two branches.

Add the pre-release indicator "ea" and the build number to the version
strings of the beta and edge channels.