Build on the core18 base to support older systems
Build on the core18 base with Ubuntu 18.04 LTS so that the JDK programs
require only GNU C Library version 2.27 instead of 2.29.

Create a chain of trusted builds that starts with Ubuntu OpenJDK
11. Publish the builds of OpenJDK 12, 13, and 14 in a new Snap package
called BootJDK on its candidate, beta, and edge channels.

Add the i386 architecture, which wasn't available for the core20 base.

Fix the following error message when running a Java Swing application:

  Fontconfig error: Cannot load config file from /etc/fonts/fonts.conf