Add support for parallel installs
The environment variables SNAP and SNAP_COMMON might not point to
the actual locations on the system outside of the Snap container. The
actual locations contain the Snap instance name instead of the Snap
name specified in the 'snapcraft.yaml' file. With a parallel install,
the Snap name and Snap instance name are different.

For example, on Debian systems we have:

  SNAP variable   = /snap/${SNAP_NAME}/${SNAP_REVISION}
  System location = /snap/${SNAP_INSTANCE_NAME}/${SNAP_REVISION}

On Debian and Fedora systems we have:

  SNAP_COMMON variable = /var/snap/${SNAP_NAME}/common
  System location      = /var/snap/${SNAP_INSTANCE_NAME}/common

Instead of using SNAP and SNAP_COMMON, construct the system locations
using SNAP_INSTANCE_NAME. Use these system locations when setting
the JAVA_HOME and MANPATH environment variables and when printing
the location of the 'openjdk.env' environment file.