Change the version to 15.0.3+0 but don't publish
Change the 'snapcraft.yaml' file for Git tag 'jdk-15.0.3+0' but don't
publish the release to the candidate channel just yet. The version
reported by the 'java' command is still 15.0.2, so wait until the
commit for 8262541, "Bump update version for OpenJDK: jdk-15.0.3"
is included in a tagged release before publishing the update to the
candidate channel.

Add the option '--with-source-date=version' to enable reproducible
builds when the feature is fully functional.