Set license value using a simple SPDX expression
Neither the Ubuntu Software Snap package (snap-store) nor the GNOME
Software Debian package (gnome-software) understand the compound SPDX
expression '(GPL-2.0 WITH Classpath-exception-2.0)' and label it a
"Proprietary" license.[1]

Replace the compound expression with the equivalent, but deprecated,
simple SPDX expression 'GPL-2.0-with-classpath-exception', which both
software store applications recognize as a "Free" license.[2]

[1] Proprietary Software

This means that the software is owned by an individual or a
company. There are often restrictions on its use and its source code
cannot usually be accessed.

More information
https://en.wikipedia.org/wiki/Proprietary_software

[2] Free Software

This means that the software can be freely run, copied, distributed,
studied and modified.

Users are bound by the following license:

GPL-2.0-with-classpath-exception
http://spdx.org/licenses/GPL-2.0-with-classpath-exception