Upgrade to core20 base and remove i386 build
Upgrade to the core20 base for gcc version 10. Building with earlier
versions of gcc prints the warning:

  WARNING: You are using gcc 7.5.0 which is older than 10.0. This is
  not a supported configuration.

Remove the build for the i386 architecture, which is not available
in the core20 base, anyway. Building for i386 prints the warning:

  WARNING: The 32-bit x86 port is deprecated and may be removed in
  a future release.

Remove the connection to the OpenJFX Snap, which is built with a
different base and is no longer compatible with the OpenJDK Snap.

Remove the 'jpackage' command, which adds too much complexity to the
build and only ever worked for Debian packages. Use 'jlink' instead.